@W-13569661@: move OccurrenceInfo to rules.ops subpackage
